refactoring: Extract module `fuzz_config`
